+++ lib/LibCrafts-1.0/Locales/esMX-Professions.lua
@@ -1,5 +1,5 @@
 local lib = LibCraftsGetLibrary()
-local name, locale, version = "Locales-Professions", "esMX", 3
+local name, locale, version = "Locales-Professions", "esMX", 4
 local module = --[[---@type LcLocaleModule]] lib:RegisterLocaleModule(name, locale, version)
 if not module then return end
 
@@ -14,3 +14,8 @@ L["Leatherworking"] = "Peletería"
 L["Mining"] = "Minería"
 L["Poisons"] = "Venenos"
 L["Tailoring"] = "Sastrería"
+
+if lib.env.is_turtle_wow then
+    L["Disguise"] = "Disfarce"
+    L["Jewelcrafting"] = "Joalheria"
+end