+++ lib/LibCrafts-1.0/Locales/koKR-Professions.lua
@@ -1,5 +1,5 @@
 local lib = LibCraftsGetLibrary()
-local name, locale, version = "Locales-Professions", "koKR", 4
+local name, locale, version = "Locales-Professions", "koKR", 5
 local module = --[[---@type LcLocaleModule]] lib:RegisterLocaleModule(name, locale, version)
 if not module then return end
 
@@ -16,5 +16,6 @@ L["Poisons"] = "독 조제"
 L["Tailoring"] = "재봉술"
 
 if lib.env.is_turtle_wow then
+    L["Disguise"] = "변장술"
     L["Jewelcrafting"] = "보석세공"
 end