+++ lib/LibCrafts-1.0/Locales/ptBR-Professions.lua
@@ -1,5 +1,5 @@
 local lib = LibCraftsGetLibrary()
-local name, locale, version = "Locales-Professions", "ptBR", 4
+local name, locale, version = "Locales-Professions", "ptBR", 5
 local module = --[[---@type LcLocaleModule]] lib:RegisterLocaleModule(name, locale, version)
 if not module then return end
 
@@ -16,6 +16,7 @@ L["Poisons"] = "Venenos"
 L["Tailoring"] = "Alfaiataria"
 
 if lib.env.is_turtle_wow then
+    L["Disguise"] = "Disfarce"
     L["Enchanting"] = "Encantador"
     L["First Aid"] = "Primeiro socorro"
     L["Jewelcrafting"] = "Joalheria"